Refactor: move workflow card to separate component
